requirements.txt
pandas==2.2.3
pyarrow==17.0.0
requests==2.32.3
python-dotenv==1.0.1
SQLAlchemy==2.0.35
psycopg[binary]==3.2.9
duckdb==1.1.3
tqdm==4.66.5
typer==0.12.5
loguru==0.7.2